requirements.txt
blinker==1.4
click==6.7
decorator==4.1.2
Flask==0.12.2
Flask-SQLAlchemy==2.2
Flask-WhooshAlchemy==0.56
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
pbr==3.1.1
pkg-resources==0.0.0
six==1.10.0
SQLAlchemy==1.1.12
sqlalchemy-migrate==0.11.0
sqlparse==0.2.3
Tempita==0.5.2
Werkzeug==0.12.2
Whoosh==2.7.4
